refactor: first version
