--- dubbogo/client/conf/dubbogo.yml
@@ -1,14 +1,0 @@
-dubbo:
-  registries:
-    demoZK:
-      protocol: zookeeper
-      address: 127.0.0.1:2181
-  protocols:
-    dubbo:
-      name: dubbo
-      port: 20000
-  consumer:
-    references:
-      UserProvider:
-        protocol: dubbo
-        interface: org.apache.dubbo.UserProvider
