--- dubbogo/server/conf/dubbogo.yml
@@ -1,13 +1,0 @@
-dubbo:
-  registries:
-    demoZK:
-      protocol: zookeeper
-      address: 127.0.0.1:2181
-  protocols:
-    dubbo:
-      name: dubbo
-      port: 20000
-  provider:
-    services:
-      UserProvider:
-        interface: org.apache.dubbo.UserProvider
