--- pixiu/dubbo/go-server/conf/dubbogo.yml
@@ -1,34 +1,0 @@
-# dubbo server yaml configure file
-
-dubbo:
-  application: # 应用信息，服务启动后会将相关信息注册到注册中心，可被客户端从 url 中识别
-    name: GreetProvider # application=myApp; name=myApp
-    group: dubbo-test # organization=dubbo
-    version: myversion # app.version=myversion
-  registries:
-    demoZK:
-      protocol: zookeeper
-      timeout: 3s
-      address: node1:2181
-  protocols:
-    dubbo:
-      name: dubbo
-      port: 20004
-  provider:
-    services:
-      UserProvider:
-        interface: org.apache.dubbo.sample.UserProvider
-#      UserProvider1:
-#        interface: org.apache.dubbo.sample.UserProvider
-#        version: "2.0"
-#      UserProvider2:
-#        interface: org.apache.dubbo.sample.UserProvider
-#        version: "2.0"
-#        group: as
-#      ComplexProvider:
-#        interface: org.apache.dubbo.sample.ComplexProvider
-#      WrapperArrayClassProvider:
-#        interface: org.apache.dubbo.sample.WrapperArrayClassProvider
-  logger:
-    zap-config:
-      level: info
