--- pixiu/triple/go-client/conf/dubbogo.yml
@@ -1,11 +1,0 @@
-dubbo:
-  registries:
-    demoZK:
-      protocol: zookeeper
-      timeout: 3s
-      address: yuluo-13:2181
-  consumer:
-    references:
-      GreeterClientImpl:
-        protocol: tri
-        interface: "org.apache.dubbogo.samples.api.Greeter" # read interface from pb
