+++ protocol/dubbo/go-client/conf/dubbogo.yml
@@ -1,0 +1,15 @@
+# dubbo client yaml configure file
+
+dubbo:
+  registries:
+    demoZK:
+      protocol: zookeeper
+      address: 127.0.0.1:2181
+  consumer:
+    references:
+      UserProvider:
+        protocol: dubbo
+        interface: org.apache.dubbo.sample.UserProvider
+  logger:
+    zap-config:
+      level: info
