+++ protocol/dubbo/go-server/cmd/server.go
@@ -1,0 +1,60 @@
+package main
+
+import (
+	"fmt"
+	"os"
+	"os/signal"
+	"syscall"
+	"time"
+)
+
+import (
+	"dubbo.apache.org/dubbo-go/v3/config"
+	_ "dubbo.apache.org/dubbo-go/v3/imports"
+
+	hessian "github.com/apache/dubbo-go-hessian2"
+
+	"github.com/dubbogo/gost/log/logger"
+)
+
+import (
+	"dubbo-go-pixiu-benchmark/protocol/dubbo/go-server/pkg"
+)
+
+func main() {
+	// ------for hessian2------
+	hessian.RegisterJavaEnum(pkg.Gender(pkg.MAN))
+	hessian.RegisterJavaEnum(pkg.Gender(pkg.WOMAN))
+	hessian.RegisterPOJO(&pkg.User{})
+	config.SetProviderService(&pkg.UserProvider{})
+	// ------------
+
+	if err := config.Load(config.WithPath("/mnt/d/Workspace/benchmark/protocol/dubbo/go-server/conf/dubbogo.yml")); err != nil {
+		panic(err)
+	}
+
+	initSignal()
+}
+
+func initSignal() {
+	signals := make(chan os.Signal, 1)
+	// It is not possible to block SIGKILL or syscall.SIGSTOP
+	signal.Notify(signals, os.Interrupt, syscall.SIGHUP, syscall.SIGQUIT, syscall.SIGTERM, syscall.SIGINT)
+	for {
+		sig := <-signals
+		logger.Infof("get signal %s", sig.String())
+		switch sig {
+		case syscall.SIGHUP:
+			// reload()
+		default:
+			time.AfterFunc(time.Duration(int(3e9)), func() {
+				logger.Warnf("app exit now by force...")
+				os.Exit(1)
+			})
+
+			// The program exits normally or timeout forcibly exits.
+			fmt.Println("provider app exit now...")
+			return
+		}
+	}
+}
