+++ protocol/dubbo/go-server/conf/dubbogo.yml
@@ -1,0 +1,19 @@
+# dubbo server yaml configure file
+
+dubbo:
+  registries:
+    demoZK:
+      protocol: zookeeper
+      timeout: 3s
+      address: 127.0.0.1:2181
+  protocols:
+    dubbo:
+      name: dubbo
+      port: 20000
+  provider:
+    services:
+      UserProvider:
+        interface: org.apache.dubbo.sample.UserProvider
+  logger:
+    zap-config:
+      level: info
