--- test/tester.go
@@ -1,42 +1,0 @@
-package test
-
-import (
-	"reflect"
-	"strings"
-	"testing"
-)
-
-type Tester struct{}
-
-func RunSubTests(t *testing.T, x interface{}) {
-	xt := reflect.TypeOf(x)
-	xv := reflect.ValueOf(x)
-
-	setup := getTestFunc(t, xv, "Setup")
-	teardown := getTestFunc(t, xv, "Teardown")
-
-	for i := 0; i < xt.NumMethod(); i++ {
-		methodName := xt.Method(i).Name
-		if !strings.HasPrefix(methodName, "Test") {
-			continue
-		}
-		tfunc := getTestFunc(t, xv, methodName)
-		t.Run(strings.TrimPrefix(methodName, "Test"), func(t *testing.T) {
-
-			t.Cleanup(func() { teardown(t) })
-			setup(t)
-			tfunc(t)
-		})
-	}
-}
-
-func getTestFunc(t *testing.T, xv reflect.Value, name string) func(*testing.T) {
-	if m := xv.MethodByName(name); m.IsValid() {
-		if f, ok := m.Interface().(func(*testing.T)); ok {
-			return f
-		}
-		// Method exists but has the wrong type signature.
-		t.Fatalf("grpctest: function %v has unexpected signature (%T)", name, m.Interface())
-	}
-	return func(*testing.T) {}
-}
